Added a theory for sets of ordered things
Builds on balanced_mapTheory.

An 'a oset is just a balanced map from 'a to unit.

There is a good_oset predicate (since not everrything in the balanced
tree type is actually a balanced binary search tree).

The following functions are defined as abbreviations over balanced
operations. They are prefix and all take a comparison operator.

oset	: pred_set
-------------------
oin 	: IN
oinsert : INSERT
odelete : DELETE
ounion 	: UNION
oimage	: IMAGE

Theorems are proved that each creates good osets, and that specifies how
membership interacts with insert, delete, union, and image (although the
image theorems are currently cheated).