Sketch out 1 case of how the conservativity proof should work.
It will need a lemma proving that the inference system respects the
quotient induced by the alphaordering used to build the sets. NB, this
isn't just to do proper up-to-alpha-equivalence reasoning. It is also to
handle the fact that osets can have many structurally different trees
representing the same set.