Further progress on parameterised relation
The idea seems to be working, but there is still more to do before I'm
certain. The basic idea is that the logical relation is parameterised by
a "world". The world maps locations (as would appear in an App SOME) to
location#arity#expression tuples. Closures with SOME annotations are
only related when the worlds expression is related to the closure's
expression, if the location of the SOME is present in the world
(otherwise there is no restriction). Furthermore, the code table must
hold an expression at the location of the world's range that is related
to the world's expression.